requirements/requirements.txt
asgiref==3.6.0
black==23.1.0
bleach==6.0.0
click==8.1.3
crispy-bootstrap5==0.7
Django==4.1.5
django-crispy-forms==1.14.0
django-filter==23.2
django-webpack-loader==1.8.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
importlib-metadata==6.6.0
gunicorn==20.1.0
h11==0.14.0
isort==5.12.0
Markdown==3.4.3
mypy-extensions==0.4.3
odict==1.9.0
packaging==23.0
pathspec==0.11.0
pillow==10.4.0
platformdirs==2.6.2
plumber==1.7
psycopg2-binary==2.9.5
pytz==2023.3
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.7
uvicorn==0.22.0
webencodings==0.5.1
zipp==3.15.0
zope.component==5.1.0
zope.deferredimport==4.4
zope.deprecation==4.4.0
zope.event==4.6
zope.hookable==5.4
zope.interface==5.5.2
zope.lifecycleevent==4.4
zope.proxy==5.0.0
